packages/type-r-adjuster/pyproject.toml
[project]
name = "type-r-adjuster"
version = "0.1.0"
description = "Type-R adjuster package."
readme = "README.md"
authors = [
    { name = "Wataru Shimoda", email = "wataru_shimoda@cyberagent.co.jp" },
]
requires-python = ">=3.10, <3.13"
dependencies = [
    "levenshtein>=0.27.1",
    "loguru>=0.7.3",
    "numpy>=2.2.3",
    "openai>=1.65.5",
    "pillow>=10.4.0",
    "pydantic>=2.10.6",
    "skia-python>=87.6",
    "type-r-core",
    "type-r-eraser",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["type_r"]  # src layout does not work in editable mode.

[tool.uv.sources]
type-r-core = { workspace = true }
type-r-eraser = { workspace = true }